# expression: implement vectorized evaluation for builtinUTCDateSig